+++ config/routes.rb
@@ -45,7 +45,7 @@ Rails.application.routes.draw do
           get '/history', to: :history, as: 'history'
           match '/custom_action', to: :custom_action, as: 'member_action', via: :all
           get '/', to: :show, as: 'show'
-          put '/', to: :update, as: 'update'
+          patch '/', to: :update, as: 'update'
           delete '/', to: :destroy, as: 'destroy'
         end