+++ lib/generators/ab_admin/install/templates/models/gender_type.rb
@@ -1,5 +1,5 @@
 class GenderType < AbAdmin::Models::TypeModel
-  self.codes = [:male, :female]
+  self.codes = [:male, :female, :undefined]
   self.i18n_scope = [:admin, :gender]
 
   define_enum_by_codes