+++ spec/dummy/config/initializers/ab_admin.rb
@@ -9,7 +9,23 @@ if Object.const_defined?('AbAdmin')
 end
 
 I18n.available_locales = Globalize.available_locales = [:ru, :en]
+I18n.enforce_available_locales = false
 
 require 'factory_girl'
 FactoryGirl.definition_file_paths = [File.expand_path('../../factories/', __FILE__)]
 FactoryGirl.find_definitions
+
+if Rails.env.development?
+  require 'sass'
+  require 'sass/engine'
+
+  module Sass
+    class Engine
+      def initialize(template, options={})
+        @options = self.class.normalize_options(options)
+        @options[:debug_info] = true
+        @template = template
+      end
+    end
+  end
+end